refac 1 more, pages/index has everything
