+++ src/pages/Asteroids/index.ts
@@ -1,7 +1,7 @@
 import GameStart from "./GameStart.png";
 import Level5 from "./Level5.png";
 import GameOver from "./GameOver.png";
-import { Page } from "../types";
+import { Page } from "../";
 
 export default {
   id: "asteroids",
