+++ src/pages/ML/index.ts
@@ -1,4 +1,4 @@
-import { Page } from "../types";
+import { Page } from "../";
 
 export default {
   id: "ml-algos",
