+++ src/pages/Sahara/index.ts
@@ -7,7 +7,7 @@ import MilHomeFAQs from "./MilHomeFAQs.png";
 import MilAboutReg from "./MilAboutReg.png";
 import SaharaHomeWork from "./SaharaHomeWork.png";
 import SaharaSidenavBlogs from "./SaharaSidenavBlogs.png";
-import { Page } from "../types";
+import { Page } from "../";
 
 export default {
   id: "sahara",
